removed pretty complains bc i can

+++ src/util/pretty.hpp
@@ -5,8 +5,8 @@
 namespace Clockwork {
 
 void print_progress(size_t current, size_t total, size_t bar_width = 40) {
-    f32    progress = static_cast<float>(current) / total;
-    size_t pos      = static_cast<size_t>(bar_width * progress);
+    f32    progress = static_cast<f32>(current) / static_cast<f32>(total);
+    size_t pos      = bar_width * static_cast<size_t>(progress);
 
     std::cout << "\r[";
     for (size_t i = 0; i < bar_width; ++i) {
